CAMEL-20843: further split the CamelTestSupport responsibilities
- Move the legacy test/context setup methods to the AbstractTestSupport
- Moved most of the helper logic to the TestSupport
- Cleaned up one affected test
- Added documentation and comments
- Deprecated a few more methods
