+++ components/camel-pubnub/src/test/java/org/apache/camel/component/pubnub/PubNubTestBase.java
@@ -39,11 +39,13 @@ public class PubNubTestBase extends CamelTestSupport {
 
     private WireMockServer wireMockServer = new WireMockServer(options().port(port));
 
+    @Override
     protected void setupResources() {
         wireMockServer.start();
         WireMock.configureFor("localhost", wireMockServer.port());
     }
 
+    @Override
     protected void cleanupResources() {
         wireMockServer.stop();
         pubnub.destroy();
