+++ components/camel-test/camel-test-junit5/src/main/java/org/apache/camel/test/junit5/CommonTestSupport.java
@@ -1,0 +1,40 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.camel.test.junit5;
+
+import org.apache.camel.CamelContext;
+
+/**
+ * A common interface for tests that use the CamelContext.
+ */
+public interface CommonTestSupport {
+
+    /**
+     * Gets the {@link CamelContextConfiguration} for the test
+     *
+     * @return the camel context configuration
+     */
+    CamelContextConfiguration camelContextConfiguration();
+
+    /**
+     * Gets the {@link TestExecutionConfiguration} test execution configuration instance for the test
+     *
+     * @return the configuration instance for the test
+     */
+    TestExecutionConfiguration testConfiguration();
+}
